Mixture of Latent Trait Analyzers

Mixture of latent trait analyzers (MLTA) has been introduced by Gollini and Murphy (2014) and Gollini (in press) to identify groups assuming the existence of a latent trait describing the dependence structure between receiver nodes within groups of sender nodes and therefore capturing the heterogeneity of sender nodes' behaviour within groups. The function mlta makes use of a variational inferential approach. For more details see Gollini, I. (in press) and Gollini, I., and Murphy, T. B. (2014).

mlta(X, G, D, wfix = FALSE, nstarts = 3, tol = 0.1^2,
  maxiter = 250, pdGH = 21)

Arguments

X
(N x M) binary incidence matrix

G
number of groups

D
dimension of the continuous latent variable

wfix
Logical. Fit the parsiomonius model with the w parameters equal across groups. Default wfix = FALSE

nstarts
number of starts. Default nstarts = 3

tol
desired tolerance for convergence. Default tol = 0.1^2

maxiter
maximum number of iterations. Default maxiter = 500

pdGH
number of quadrature points for the Gauss-Hermite quadrature. Default pdGH = 21

Value

List containing the following information for each model fitted:

b matrix containing intercepts for the logistic response function

w array containing slopes for the logistic response function

eta \(\eta_g\) is the mixing proportion for the group \(g (g = 1,..., G)\), that corresponds to the prior probability that a randomly chosen sender node is in the g-th group.

mu (N x D x G) array containing posterior means for the latent variable

C (D x D x N x G) array containing posterior variances for the latent variable

z (N x G) matrix containing posterior probability for each sender node to belong to each group

LL log likelihood

BIC Bayesian Information Criterion (BIC) (Schwarz (1978))

If multiple models are fitted the output contains also tables to compare the log likelihood and BIC for all models fitted.

References

Gollini, I. (in press) 'A mixture model approach for clustering bipartite networks', Challenges in Social Network Research Volume in the Lecture Notes in Social Networks (LNSN - Series of Springer). Preprint: https://arxiv.org/abs/1905.02659.

Gollini, I., and Murphy, T. B. (2014), 'Mixture of Latent Trait Analyzers for Model-Based Clustering of Categorical Data', Statistics and Computing, 24(4), 569-588 http://arxiv.org/abs/1301.2167.

See also

Examples

### Simulate Bipartite Network
set.seed(1)
X <- matrix(rbinom(4 * 12, size = 1, prob = 0.4), nrow = 12, ncol = 4)

resMLTA <- mlta(X, G = 2, D = 1)